+++ src/blocks/grid/edit.js
@@ -1,7 +1,4 @@
-import wasBlockJustInserted from '../../utils/was-block-just-inserted';
-import isBlockVersionLessThan from '../../utils/check-block-version';
-import hasNumericValue from '../../utils/has-numeric-value';
-import { useDispatch, useSelect } from '@wordpress/data';
+import { useDispatch } from '@wordpress/data';
 import { Fragment, useEffect, useState } from '@wordpress/element';
 import { InnerBlocks } from '@wordpress/block-editor';
 import LayoutSelector, { getColumnsFromLayout } from './components/LayoutSelector';
@@ -13,9 +10,9 @@ import MainCSS from './css/main';
 import ComponentCSS from './components/ComponentCSS';
 import classnames from 'classnames';
 import { applyFilters } from '@wordpress/hooks';
-import withUniqueId from '../../hoc/withUniqueId';
 import { compose } from '@wordpress/compose';
 import { useDeviceType, useInnerBlocksCount } from '../../hooks';
+import { withUniqueId, withGridLegacyMigration } from '../../hoc';
 
 const GridEdit = ( props ) => {
 	const {
@@ -35,34 +32,6 @@ const GridEdit = ( props ) => {
 			columns: innerBlocksCount,
 		} );
 	}, [ innerBlocksCount ] );
-
-	useEffect( () => {
-		if ( ! attributes.isDynamic ) {
-			setAttributes( { isDynamic: true } );
-		}
-
-		// Set our old defaults as static values.
-		// @since 1.4.0.
-		if ( ! wasBlockJustInserted( attributes ) && isBlockVersionLessThan( attributes.blockVersion, 2 ) ) {
-			const legacyDefaults = generateBlocksLegacyDefaults.v_1_4_0.gridContainer;
-
-			const newAttrs = {};
-
-			const hasGlobalStyle = attributes.useGlobalStyle && attributes.globalStyleId;
-
-			if ( ! hasGlobalStyle && ! hasNumericValue( attributes.horizontalGap ) ) {
-				newAttrs.horizontalGap = legacyDefaults.horizontalGap;
-			}
-
-			if ( Object.keys( newAttrs ).length > 0 ) {
-				setAttributes( newAttrs );
-			}
-		}
-
-		if ( isBlockVersionLessThan( attributes.blockVersion, 2 ) ) {
-			setAttributes( { blockVersion: 2 } );
-		}
-	}, [] );
 
 	useEffect( () => {
 		if ( selectedLayout ) {
@@ -129,4 +98,7 @@ const GridEdit = ( props ) => {
 	);
 };
 
-export default compose( withUniqueId )( GridEdit );
+export default compose(
+	withUniqueId,
+	withGridLegacyMigration,
+)( GridEdit );
